fix(nestjs): eslint errros correction

+++ src/modules/tasks/tasks-manager.service.ts
@@ -1,11 +1,11 @@
 import { Injectable, Logger } from '@nestjs/common';
-import { NosicaService} from './nosica/nosica.service';
+import { NosicaService } from './nosica/nosica.service';
 import { PyramidService } from './pyramid/pyramid.service';
 import { MyGtsService } from './mgts/mygts.service';
 import { ConstantService } from '../constants/constants';
 import { Cron, CronExpression } from '@nestjs/schedule';
 import { readFileSync, readdirSync } from 'fs';
-import AWS  = require('aws-sdk');
+import AWS = require('aws-sdk');
 import * as csvParser from 'csv-parser';
 import { map, includes } from 'lodash';
 import * as stringToStream from 'string-to-stream';
@@ -14,7 +14,7 @@ import path = require('path');
 const secureEnvs = ['homologation', 'production'];
 
 @Injectable()
-export class TasksService  {
+export class TasksService {
   private readonly logger = new Logger(TasksService.name);
   private S3 = new AWS.S3({ endpoint: process.env.AWS_BUCKET_ENDPOINT });
   constructor(
@@ -22,16 +22,16 @@ export class TasksService  {
     private pyramidService: PyramidService,
     private constantService: ConstantService,
     private myGtsService: MyGtsService,
-  ) {} 
+  ) {}
 
   getFlowType(flow: string): string {
-    const datas  = flow.split('.');
-    if (datas.length !== 5 )  return undefined
+    const datas = flow.split('.');
+    if (datas.length !== 5) return undefined;
     return datas[3];
   }
 
   importLine(flow, line) {
-    switch(flow) {
+    switch (flow) {
       case this.constantService.GLOBAL_CONST.QUEUE.EAC.NAME:
         return this.pyramidService.import(line);
       case this.constantService.GLOBAL_CONST.QUEUE.PMD.NAME:
@@ -48,23 +48,18 @@ export class TasksService  {
   parseFile(buffer, filename): Promise<string> {
     const readable = stringToStream(buffer.toString());
     const flowType = this.getFlowType(filename);
-    const separator = this.constantService.GLOBAL_CONST.QUEUE[flowType].ORIGIN_SEPARATOR;  
-    const _this = this;
-    const stream = readable.pipe (csvParser({ separator: separator }));
+    const separator = this.constantService.GLOBAL_CONST.QUEUE[flowType].ORIGIN_SEPARATOR;
+    const stream = readable.pipe(csvParser({ separator: separator }));
 
     return new Promise((resolve, reject) => {
-      stream.on('data', async (parsed) => {
-        stream.pause();
-        await _this.importLine(flowType, parsed)
-        .catch((error) => _this.logger.error(`${filename} error occurred: ${error}`));
-        stream.resume();
-      })
-      .on('end', () => {
-        resolve('end');
-      })
-      .on('error', (err) => {
-        reject(err);
-      })
+      stream
+        .on('data', async parsed => {
+          stream.pause();
+          await this.importLine(flowType, parsed).catch(error => this.logger.error(`${filename} error occurred: ${error}`));
+          stream.resume();
+        })
+        .on('end', () => resolve('end'))
+        .on('error', err => reject(err));
     });
   }
 
@@ -73,32 +68,29 @@ export class TasksService  {
    */
   @Cron(CronExpression.EVERY_HOUR)
   async importFromS3() {
-    if(!includes(secureEnvs,process.env.NODE_ENV))
-      return false;
+    if (!includes(secureEnvs, process.env.NODE_ENV)) return false;
 
     let params: any = {
       Bucket: `${process.env.AWS_BUCKET_PREFIX}gpc-set`,
     };
 
     const objects: any = await this.S3.listObjects(params).promise();
-    if(!objects || !objects.Contents.length) {
-      this.logger.log(`No file found in S3 GPC bucket`); 
+    if (!objects || !objects.Contents.length) {
+      this.logger.log(`No file found in S3 GPC bucket`);
       return;
     }
 
     map(objects.Contents, async file => {
       const flow = file.Key;
       const flowType = this.getFlowType(flow);
-      params = {...params,  Key: flow }
-      
-      if(!flowType) return;
-      if(!this.constantService.GLOBAL_CONST.QUEUE[flowType]) return;
-      
-      const s3object = (await this.S3.getObject(params).promise());
+      params = { ...params, Key: flow };
+
+      if (!flowType) return;
+      if (!this.constantService.GLOBAL_CONST.QUEUE[flowType]) return;
+
+      const s3object = await this.S3.getObject(params).promise();
       const parsed = await this.parseFile(s3object.Body, flow);
-      if(parsed) {
-        this.S3.deleteObject(params).promise();
-      }
+      if (parsed) this.S3.deleteObject(params).promise();
     });
   }
 
@@ -107,22 +99,19 @@ export class TasksService  {
    */
   @Cron(CronExpression.EVERY_10_SECONDS)
   importFromLocal() {
+    if (includes(secureEnvs, process.env.NODE_ENV)) return false;
 
-    if(includes(secureEnvs,process.env.NODE_ENV))
-      return false;
-    
-    const receptiondir = `${ __dirname }/../../set/reception`;
-    const files = readdirSync(receptiondir, { withFileTypes: true}).filter(file => file.isFile());
+    const receptiondir = `${__dirname}/../../set/reception`;
+    const files = readdirSync(receptiondir, { withFileTypes: true }).filter(file => file.isFile());
 
-    if(!files || !files.length) {
-      this.logger.log(`No file found in reception dir`); 
+    if (!files || !files.length) {
+      this.logger.log(`No file found in reception dir`);
       return;
     }
 
-    map(files, async(file) => {
+    map(files, async file => {
       const lines = readFileSync(path.join(receptiondir, file.name));
-      const parsed = await this.parseFile(lines, file.name);
-    });    
+      await this.parseFile(lines, file.name);
+    });
   }
-
 }
